+++ app/constants/decred.js
