+++ app/constants/index.js
@@ -1,7 +1,6 @@
-// XXX change all files to small letters!
-export * from "./Decred";
-export * from "./Decrediton";
-export * from "./Messages";
+export * from "./decred";
+export * from "./decrediton";
+export * from "./messages";
 export * from "./config";
 export * from "./opcode";
 export * from "./governance";
